AShypkov: Module 2: gRPC part: refactoring for packages
