+++ module-2-rpc-protocols-data-formats/src/main/java/com/epam/advanced/java/grpc/client/MessageExchangeClient.java
@@ -1,9 +1,9 @@
-package com.epam.advanced.java.client;
+package com.epam.advanced.java.grpc.client;
 
 import com.epam.advanced.java.exchange.MessageRequest;
 import com.epam.advanced.java.exchange.MessageResponse;
 import com.epam.advanced.java.exchange.MessagesExchangeServiceGrpc;
-import com.epam.advanced.java.utils.TimestampUtils;
+import com.epam.advanced.java.grpc.utils.TimestampUtils;
 import io.grpc.Channel;
 import io.grpc.ManagedChannelBuilder;
 import lombok.extern.slf4j.Slf4j;
