+++ module-2-rpc-protocols-data-formats/src/main/java/com/epam/advanced/java/grpc/utils/TimestampUtils.java
@@ -1,4 +1,4 @@
-package com.epam.advanced.java.utils;
+package com.epam.advanced.java.grpc.utils;
 
 import com.google.protobuf.Timestamp;
 import lombok.AccessLevel;
